Fix handling of deletion events in frontend
